Update code from Unvanquished
- Use same commit for recastnavigation submodule
- Update bot headers. These are moved to libs/unvanquished
  since the code now lives in Unvanquished, not Daemon.
