+++ libs/unvanquished/src/sgame/botlib/nav.h
@@ -1,24 +1,36 @@
 /*
 ===========================================================================
-Copyright (C) 2012 Unvanquished Developers
 
-This file is part of Daemon source code.
+Daemon BSD Source Code
+Copyright (c) 2013 Daemon Developers
+All rights reserved.
 
-Daemon source code is free software; you can redistribute it
-and/or modify it under the terms of the GNU General Public License as
-published by the Free Software Foundation; either version 2 of the License,
-or (at your option) any later version.
+This file is part of the Daemon BSD Source Code (Daemon Source Code).
 
-Daemon source code is distributed in the hope that it will be
-useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+    * Redistributions of source code must retain the above copyright
+      notice, this list of conditions and the following disclaimer.
+    * Redistributions in binary form must reproduce the above copyright
+      notice, this list of conditions and the following disclaimer in the
+      documentation and/or other materials provided with the distribution.
+    * Neither the name of the Daemon developers nor the
+      names of its contributors may be used to endorse or promote products
+      derived from this software without specific prior written permission.
 
-You should have received a copy of the GNU General Public License
-along with Daemon source code; if not, write to the Free Software
-Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL DAEMON DEVELOPERS BE LIABLE FOR ANY
+DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+
 ===========================================================================
 */
+
 // nav.h -- navigation mesh definitions
 #ifndef __NAV_H
 #define __NAV_H
@@ -225,12 +237,12 @@ struct LinearAllocator : public dtTileCacheAlloc
 	size_t top;
 	size_t high;
 
-	LinearAllocator( const size_t cap ) : buffer(0), capacity(0), top(0), high(0)
+	LinearAllocator( const size_t cap ) : buffer(nullptr), capacity(0), top(0), high(0)
 	{
 		resize(cap);
 	}
 
-	~LinearAllocator() OVERRIDE
+	~LinearAllocator() override
 	{
 		free( buffer );
 	}
@@ -246,22 +258,22 @@ struct LinearAllocator : public dtTileCacheAlloc
 		capacity = cap;
 	}
 
-	void reset() OVERRIDE
+	void reset() override
 	{
 		high = dtMax( high, top );
 		top = 0;
 	}
 
-	void* alloc( const size_t size ) OVERRIDE
+	void* alloc( const size_t size ) override
 	{
 		if ( !buffer )
 		{
-			return 0;
+			return nullptr;
 		}
 
 		if ( top + size > capacity )
 		{
-			return 0;
+			return nullptr;
 		}
 
 		unsigned char* mem = &buffer[ top ];
@@ -269,7 +281,7 @@ struct LinearAllocator : public dtTileCacheAlloc
 		return mem;
 	}
 
-	void free( void* /*ptr*/ ) OVERRIDE { }
+	void free( void* /*ptr*/ ) override { }
 	size_t getHighSize() { return high; }
 };
 #endif
